+++ internal/manifest/manifest.go
@@ -2,13 +2,13 @@ package manifest
 
 // Manifest is the representation of the codebase
 type Manifest struct {
-	Projects map[string]Project `json:"projects"`
-	Commands map[string]string  `json:"commands"`
+	Projects map[string]Project  `json:"projects"`
+	Scripts  map[string][]string `json:"scripts"`
 }
 
 // Project is a Codebase project
 type Project struct {
-	Remote   string            `json:"remote"`
-	Config   map[string]string `json:"config"`
-	Commands map[string]string `json:"commands"`
+	Remote  string              `json:"remote"`
+	Config  map[string]string   `json:"config"`
+	Scripts map[string][]string `json:"scripts"`
 }